+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Add Rtsp Stream Link
+
 ## 2.10.2 (2023-08-08)
 
 * (theimo1221) Add BlockAutomaticHandler to zigbeeDimmer and it's children